feature: refactor message routes
